+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/app/chartdata/ChartData.java
@@ -1,7 +1,11 @@
 package gov.healthit.chpl.app.chartdata;
 
+import java.util.List;
+
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+
+import gov.healthit.chpl.domain.CertifiedProductSearchDetails;
 
 /**
  * This is the starting point for populating statistics tables that will be used for the
@@ -24,8 +28,34 @@ public final class ChartData {
         ChartDataApplicationEnvironment appEnvironment = null;
         try {
             appEnvironment = new ChartDataApplicationEnvironment();
-            SedParticipantsStatisticCount sedParticipantsStatisticCount = new SedParticipantsStatisticCount();
-            sedParticipantsStatisticCount.run(appEnvironment);
+            
+            SedDataCollector sedDataCollector = new SedDataCollector();
+            
+            List<CertifiedProductSearchDetails> seds = sedDataCollector.retreiveData(appEnvironment);
+            
+            SedParticipantsStatisticCountCalculator sedParticipantsStatisticCountCalculator = 
+                    new SedParticipantsStatisticCountCalculator();
+            sedParticipantsStatisticCountCalculator.run(seds, appEnvironment);
+            
+            ParticipantGenderStatisticsCalculator participantGenderStatisticsCalculator = 
+                    new ParticipantGenderStatisticsCalculator();
+            participantGenderStatisticsCalculator.run(seds, appEnvironment);
+            
+            ParticipantAgeStatisticsCalculator participantAgeStatisticsCalculator =
+                    new ParticipantAgeStatisticsCalculator();
+            participantAgeStatisticsCalculator.run(seds, appEnvironment);
+            
+            ParticipantEducationStatisticsCalculator participantEducationStatisticsCalculator =
+                    new ParticipantEducationStatisticsCalculator();
+            participantEducationStatisticsCalculator.run(seds, appEnvironment);
+            
+            ParticipantExperienceStatisticsCalculator participantProfExperienceStatisticsCalculator =
+                    new ParticipantExperienceStatisticsCalculator();
+            
+            participantProfExperienceStatisticsCalculator.run(seds, ExperienceType.COMPUTER_EXPERIENCE, appEnvironment);
+            participantProfExperienceStatisticsCalculator.run(seds, ExperienceType.PRODUCT_EXPERIENCE, appEnvironment);
+            participantProfExperienceStatisticsCalculator.run(seds, ExperienceType.PROFESSIONAL_EXPERIENCE, appEnvironment);
+
         } catch (Exception e) {
             LOGGER.error("Fatal Error Running ChartData! " + e.getMessage(), e);
         } finally {