+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/app/chartdata/SedDataCollectorAsyncHelper.java
@@ -1,0 +1,39 @@
+package gov.healthit.chpl.app.chartdata;
+
+import java.util.concurrent.Future;
+
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+import org.springframework.scheduling.annotation.Async;
+import org.springframework.scheduling.annotation.AsyncResult;
+import org.springframework.stereotype.Service;
+
+import gov.healthit.chpl.dao.EntityRetrievalException;
+import gov.healthit.chpl.domain.CertifiedProductSearchDetails;
+import gov.healthit.chpl.manager.CertifiedProductDetailsManager;
+
+/**
+ * Provides asynchronous support to SedParticipantsStatisticsCount class for retrieving certified product details.
+ * @author TYoung
+ *
+ */
+@Service("sedDataCollectorAsyncHelper")
+public class SedDataCollectorAsyncHelper {
+    private static final Logger LOGGER = LogManager.getLogger(SedDataCollectorAsyncHelper.class);
+
+    /**
+     * Retrieves the associated CertifiedProductionSearchDetails object as a Future<>.
+     * @param id id associated to the requested CertifiedProductionSearchDetails object
+     * @param certifiedProductDetailsManager provides access to retrieve CertifiedProductionSearchDetails object
+     * @return CertifiedProductionSearchDetails object
+     * @throws EntityRetrievalException when the CertifiedProductionSearchDetails object could not be retrieved
+     */
+    @Async("chartDataExecutor")
+    public Future<CertifiedProductSearchDetails> getCertifiedProductDetail(
+            final Long id, final CertifiedProductDetailsManager certifiedProductDetailsManager)
+                throws EntityRetrievalException {
+        CertifiedProductSearchDetails dto = certifiedProductDetailsManager.getCertifiedProductDetails(id);
+        LOGGER.info("Finishing Details for: " + id);
+        return new AsyncResult<CertifiedProductSearchDetails>(dto);
+    }
+}